feat: rename networLogFormatters to networkLogUtils
