+++ packages/vscode-extension/src/network/utils/networkLogUtils.ts
